+++ tests/conftest.py
@@ -53,3 +53,4 @@ def multiple_errors_dto():
         ]
     return _create
 
+